final changes lesson6

+++ src/main/java/ru/lavrov/tm/service/TaskService.java
@@ -2,7 +2,6 @@ package ru.lavrov.tm.service;
 
 import ru.lavrov.tm.entity.Project;
 import ru.lavrov.tm.entity.Task;
-import ru.lavrov.tm.entity.User;
 import ru.lavrov.tm.exception.project.ProjectNameIsInvalidException;
 import ru.lavrov.tm.exception.project.ProjectNotExistsException;
 import ru.lavrov.tm.exception.task.TaskNameExistsException;
@@ -30,12 +29,27 @@ public class TaskService {
         this.userRepository = userRepository;
     }
 
-    public void persist(String taskName) throws RuntimeException {
+    public void createByName(String taskName, String projectName, String userId) throws RuntimeException {
         if (taskName == null || taskName.isEmpty())
             throw new TaskNameIsInvalidException();
-        if (taskRepository.findTaskByName(taskName) != null)
+        if (projectName == null || projectName.isEmpty())
+            throw new ProjectNameIsInvalidException();
+        if (userId == null)
+            throw new UserIsNotAuthorizedException();
+        Project project = projectRepository.findProjectByName(projectName);
+        if (project == null)
+            throw new ProjectNotExistsException();
+        if (!project.getUserId().equals(userId))
+            throw new ProjectNotExistsException();
+        if (taskRepository.findProjectTaskByName(taskName, project.getId(), userId) != null)
             throw new TaskNameExistsException();
-        taskRepository.persist(new Task(taskName));
+        persist(new Task(taskName, project.getId(), userId));
+    }
+
+    public void persist(Task task) throws RuntimeException {
+        if (task == null)
+            throw new TaskNotExistsException();
+        taskRepository.persist(task);
     }
 
     public void merge(String taskName) throws RuntimeException {
@@ -52,80 +66,40 @@ public class TaskService {
         return taskRepository.findAll();
     }
 
-    public Collection<Task> findAllByUser(User sessionUser){
-        if (sessionUser == null)
+    public Collection<Task> findAllByUser(String userId){
+        if (userId == null)
             throw new UserIsNotAuthorizedException();
-        return taskRepository.findAllByUser(sessionUser);
+        return taskRepository.findAllByUser(userId);
     }
 
-    public void removeTask(String taskName) throws RuntimeException {
+    public void removeTaskByName(String taskName, String userId) throws RuntimeException {
         if (taskName == null || taskName.isEmpty())
             throw new TaskNameIsInvalidException();
+        if (userId == null)
+            throw new UserIsNotAuthorizedException();
         Task task = taskRepository.findTaskByName(taskName);
         if (task == null)
+            throw new TaskNotExistsException();
+        if (!task.getUserId().equals(userId))
             throw new TaskNotExistsException();
         taskRepository.remove(task.getId());
-
     }
 
-    public void attachTaskToProject(String taskName, String projectName) throws RuntimeException {
-        if (taskName == null || taskName.isEmpty())
-            throw new TaskNameIsInvalidException();
+    public void renameTask(String projectName, String oldName, String newName, String userId) throws RuntimeException {
         if (projectName == null || projectName.isEmpty())
             throw new ProjectNameIsInvalidException();
-        Task task = taskRepository.findTaskByName(taskName);
-        if (task == null)
-            throw new TaskNotExistsException();
-        Project project = projectRepository.findProjectByName(projectName);
-        if (project == null)
-            throw new ProjectNotExistsException();
-        taskRepository.attachTaskToProject(task, project);
-    }
-
-    public void attachTaskToUser(String taskName, User sessionUser) throws RuntimeException {
-        if (taskName == null || taskName.isEmpty())
-            throw new TaskNameIsInvalidException();
-        if (sessionUser == null)
-            throw new UserIsNotAuthorizedException();
-        Task task = taskRepository.findTaskByName(taskName);
-        if (task == null)
-            throw new TaskNotExistsException();
-        taskRepository.attachTaskToUser(task, sessionUser);
-    }
-
-    public void attachTaskToUser(String taskName, String login) throws RuntimeException {
-        User user = userRepository.findUserByLogin(login);
-        attachTaskToUser(taskName, user);
-    }
-
-    public void detachTaskfromUser(User sessionUser, String taskName){
-        if (sessionUser == null)
-            throw new UserIsNotAuthorizedException();
-        if (taskName == null || taskName.isEmpty())
-            throw new TaskNameIsInvalidException();
-        Task task = taskRepository.findTaskByName(taskName);
-        if (task == null)
-            throw new TaskNotExistsException();
-        task.setUserId(null);
-    }
-
-    public void detachTaskfromUser(String login, String taskName){
-        User user = userRepository.findUserByLogin(login);
-        detachTaskfromUser(user, taskName);
-    }
-
-    public void detachTaskfromProject(String taskName){
-        if (taskName == null || taskName.isEmpty())
-            throw new TaskNameIsInvalidException();
-        Task task = taskRepository.findTaskByName(taskName);
-        if (task == null)
-            throw new TaskNotExistsException();
-        task.setProjectId(null);
-    }
-
-    public void renameTask(String oldName, String newName) throws RuntimeException {
         if (newName == null || newName.isEmpty() || oldName == null || oldName.isEmpty())
             throw new TaskNameIsInvalidException();
-        taskRepository.renameTask(oldName, newName);
+        if (userId == null)
+            throw new UserIsNotAuthorizedException();
+        Project project = projectRepository.findProjectByName(newName);
+        if (project == null)
+            throw new ProjectNotExistsException();
+        Task task = taskRepository.findTaskByName(oldName);
+        if (task == null)
+            throw new TaskNotExistsException();
+        if (taskRepository.findProjectTaskByName(newName, project.getId(), userId) != null)
+            throw new TaskNotExistsException();
+        taskRepository.renameTask(task.getId(), newName);
     }
 }
